requirements.txt
geopy==2.2.0
google-cloud-bigquery
python-telegram-bot==13.12
